% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pids.R
\name{post_pids}
\alias{post_pids}
\title{POST PIDsvc XML Backup to PIDsvc}
\usage{
post_pids(in_f, user, password, root = "https://geoconnex.us")
}
\arguments{
\item{in_f}{character path to input xml file}

\item{user}{character username for PID service}

\item{password}{character password for PID service}

\item{root}{character URI of PID service root}
}
\value{
returns PIDsvc HTTP POST response
}
\description{
Given an input file xml, PID server URL and authentication 
will upload will upload mappings to a PID Service.
}
