% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retire_pid.R
\name{retire_pid}
\alias{retire_pid}
\title{POST PIDsvc delete_mapping request}
\usage{
retire_pid(pid, user, password, root = "https://geoconnex.us")
}
\arguments{
\item{pid}{character string for the full pid to be retired}

\item{user}{character username for PID service}

\item{password}{character password for PID service}

\item{root}{character URI of PID service root}
}
\value{
returns PIDsvc HTTP POST response
}
\description{
Given a PID, PID server URL and authentication
will retire PID in a PID Service.
}
